pyproject.toml
[project]
name = "open-rag-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.0",
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "docx2txt>=0.9",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "joblib>=1.5.3",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "langchain-unstructured>=1.0.1",
    "pdf2image>=1.17.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "python-pptx>=1.0.2",
    "tavily>=1.1.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "unstructured>=0.18.24",
]
